# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/projects:
    get:
      operationId: listProjects
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProjectsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: List all registered projects (active + degraded)
      tags:
      - projects
    post:
      operationId: addProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Register a new project from a git repository path
      tags:
      - projects
  /api/v1/projects/{id}:
    delete:
      operationId: removeProject
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveProjectResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Remove a project; stops sessions, cleans workspaces, unregisters
      tags:
      - projects
    get:
      operationId: getProject
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectGetResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Fetch one project; discriminates ok vs degraded
      tags:
      - projects
  /api/v1/projects/{id}/admission:
    get:
      operationId: getProjectAdmission
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectAdmissionState'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Read the pause on new launches; existing sessions may still be running
      tags:
      - projects
    put:
      operationId: setProjectAdmission
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetProjectAdmissionRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectAdmissionState'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Pause or allow new launches without stopping existing sessions or transferring
        ownership
      tags:
      - projects
  /api/v1/projects/{id}/config:
    put:
      operationId: setProjectConfig
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetProjectConfigInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Replace a project's per-project config
      tags:
      - projects
  /api/v1/projects/initialize:
    post:
      operationId: initializeProjectRepository
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitializeRepositoryInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitializeRepositoryResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Initialize a selected folder as a Git repository with an initial commit
      tags:
      - projects
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
    AddProjectInput:
      properties:
        asWorkspace:
          type: boolean
        config:
          $ref: '#/components/schemas/ProjectConfig'
        name:
          type:
          - "null"
          - string
        path:
          type: string
        projectId:
          type:
          - "null"
          - string
      required:
      - path
      type: object
    AgentConfig:
      properties:
        model:
          type: string
        permissions:
          type: string
        profile:
          type: string
        reasoningEffort:
          type: string
      type: object
...
    DegradedProject:
      properties:
        id:
          type: string
        kind:
          type: string
        name:
          type: string
        path:
          type: string
        resolveError:
          type: string
      required:
      - id
      - name
      - kind
      - path
      - resolveError
      type: object
...
    DomainReviewerConfig:
      properties:
        harness:
          type: string
      required:
      - harness
      type: object
...
    InitializeRepositoryInput:
      properties:
        path:
          type: string
      required:
      - path
      type: object
    InitializeRepositoryResult:
      properties:
        path:
          type: string
      required:
      - path
      type: object
...
    ListProjectsResponse:
      properties:
        projects:
          items:
            $ref: '#/components/schemas/ProjectSummary'
          type: array
      required:
      - projects
      type: object
...
    Project:
      properties:
        agent:
          type: string
        config:
          $ref: '#/components/schemas/ProjectConfig'
        defaultBranch:
          type: string
        id:
          type: string
        kind:
          type: string
        name:
          type: string
        path:
          type: string
        repo:
          type: string
        workspaceRepos:
          items:
            $ref: '#/components/schemas/WorkspaceRepo'
          type: array
      required:
      - id
      - name
      - kind
      - path
      - repo
      - defaultBranch
      type: object
    ProjectAdmissionState:
      properties:
        admissionPaused:
          type: boolean
        existingSessionsMayBeRunning:
          type: boolean
        projectId:
          type: string
        scope:
          enum:
          - new_launches_only
          type: string
      required:
      - projectId
      - admissionPaused
      - scope
      - existingSessionsMayBeRunning
      type: object
    ProjectConfig:
      properties:
        admissionPaused:
          type: boolean
        agentConfig:
          $ref: '#/components/schemas/AgentConfig'
        defaultBranch:
          type: string
        env:
          additionalProperties:
            type: string
          type: object
        orchestrator:
          $ref: '#/components/schemas/RoleOverride'
        postCreate:
          items:
            type: string
          type: array
        reviewers:
          items:
            $ref: '#/components/schemas/DomainReviewerConfig'
          type: array
        sessionPrefix:
          type: string
        symlinks:
          items:
            type: string
          type: array
        trackerIntake:
          $ref: '#/components/schemas/TrackerIntakeConfig'
        worker:
          $ref: '#/components/schemas/RoleOverride'
      type: object
    ProjectGetResponse:
      properties:
        project:
          $ref: '#/components/schemas/ProjectOrDegraded'
        status:
          enum:
          - ok
          - degraded
          type: string
      required:
      - status
      - project
      type: object
    ProjectOrDegraded:
      oneOf:
      - $ref: '#/components/schemas/Project'
      - $ref: '#/components/schemas/DegradedProject'
      type: object
    ProjectResponse:
      properties:
        project:
          $ref: '#/components/schemas/Project'
      required:
      - project
      type: object
    ProjectSummary:
      properties:
        id:
          type: string
        kind:
          type: string
        name:
          type: string
        orchestratorAgent:
          type: string
        path:
          type: string
        resolveError:
          type: string
        sessionPrefix:
          type: string
      required:
      - id
      - name
      - path
      - kind
      - sessionPrefix
      type: object
    RemoveProjectResult:
      properties:
        projectId:
          type: string
        removedStorageDir:
          type: boolean
      required:
      - projectId
      - removedStorageDir
      type: object
...
    RoleOverride:
      properties:
        agent:
          type: string
        agentConfig:
          $ref: '#/components/schemas/AgentConfig'
      type: object
...
    SetProjectAdmissionRequest:
      properties:
        paused:
          description: Whether new launches for this project are paused. Existing
            sessions may still be running.
          type: boolean
      required:
      - paused
      type: object
    SetProjectConfigInput:
      properties:
        config:
          $ref: '#/components/schemas/ProjectConfig'
      required:
      - config
      type: object
...
    TrackerIntakeConfig:
      properties:
        assignee:
          type: string
        enabled:
          type: boolean
        provider:
          enum:
          - github
          type: string
        repo:
          type: string
      type: object
...
    WorkspaceRepo:
      properties:
        name:
          type: string
        relativePath:
          type: string
        repo:
          type: string
      required:
      - name
      - relativePath
      - repo
      type: object
...